+++ NEWS.md
@@ -1,3 +1,10 @@
+v0.12.0 (2018-??-??)
+====================
+
+- Column names in FITS files are now case sensitive by default.  Functions
+  `read(hdu, colname)` and `Libcfitsio.fits_get_colnum` have a new optional
+  keyword, `case_sensitive`, which is `true` by default.
+
 v0.11.0 (2017-09-28)
 ====================